requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
SQLAlchemy==2.0.23
pandas==2.2.2
mysql-connector-python==9.0.0
xgboost==2.1.1
python-dotenv==1.0.1
gunicorn==22.0.0
